Add entropy to the kernel RNG using the rtc clock.
On QEMU, there is no entropy in the kernel RNG due to MG-384. This CL
uses the rtc clock to add some entropy.

This means that even when there is no hardware entropy generator, the
ledger is not always generating the same ids.

It is still an issue from a security perspective, as the entropy is
very small.

LE-105 #fixed

Change-Id: Ic1657a169a2571a7217336efc51ba2f90bf9ee57