Synchronize starting of app under trace.
trace record only launches the app specified on the command line
only after the TraceManager has acknowledged the Start request. With
that, we give already registered trace providers a chance to prepare and
make sure that trace data from the app under tracing can be interpreted
within the context of the other providers' data.

Change-Id: I3a47f7fc18f57bb3d7f7e0c64a011fc3c8722339