[bootloader][docs] Add a commandline param to set default boot
bootloader.default=local will default to a local magenta.bin
bootloader.default=network will default to netboot, and is the default
if this cmdline parameter is not set.

bootloader.timeout=0 is now fixed as well, and uses the default.

Also added documentation on gigaboot commandline params to magenta
docs.

MG-389 #done

Change-Id: Ib94007d3451f588e4a2d5e0ac145826a927de704