apps/media: prevent TaskRunner calling a deleted FidlPacketProducer
FidlPacketProducer uses a task runner to ensure that packets are
produced on the correct thread. On rare occasions, the producer is
destroyed with a task pending. When the task runs, it tries to use
a deleted 'this', and badness ensues. This CL provides the lambda
for the pending task with a weak_ptr to this so the task can die
quickly if the producer is gone.

Change-Id: Ia0fc18c36defc1745226c8ff89c156f8c3d54390