[devmgr] Unmount FS servers, even when they don't support IOCTL_DEVMGR_UNMOUNT_FS
This is currently an issue for the netstack, which registers itself as a
filesystem server using IOCTL_DEVMGR_MOUNT_FS. However, when the
corresponding unmount signal is sent to the netstack, it returns an
error, rather than closing the reply handle sent from txn_unmount.

This patch attempts to generalize the unmounting process slightly.
Although we are still forcefully unmounting, consider any of the
following to be an 'ack' from the filesystem server:
 - Closing the input reply handle
 - Writing an error response on the reply handle
 - Writing a success value on the reply handle

Change-Id: I66867528388a9b21aa8b657d0e7655c3ca80c3aa