Change DeviceShell from Initialize() to a factory method.
Also, change the name of the DeviceRunner service to align better with
what we call UserRunner: The UserRunner is what services above it call
when they run the user_runner application, while DeviceRunner was what
is passed to services that the device_runner application starts (specifcally,
it was passed to the DeviceShell). The analog that is passed to UserShell
is called StoryProvider. Hence, we rename the DeviceRunner service
to UserProvider. There is no DeviceRunner service, because the device_runner
application is started from the command line.

Although the name can be justified, it's a little awkward because
very little in terms of a user seems to be provided by the UserProvider
(conceivably it could be a set of users that may log in for the device
shell to select from, just as the StoryProvider gives the UserShell
a set of stories to select from). So maybe I change it further in the
future, but the different roles of DeviceRunner and UserRunner were
totally distracting to me.

Change-Id: Ie740865aad5d315aca75ffe0ce07cfe0d7b6e6f3