[ulib][musl] Declare pthread_attr_setstack as deprecated, and remove it
We'd like not to support explicit allocation of stacks, to facilitate
both enforced randomized placement and -fsanitize=safe-stack mode.
For now, just remove the entry point so we can catch any higher-level
code that is attempting to use it.  Later we can simplify and clean up
the thread-creation internals without this option.

The declaration is left behind so there is a more helpful error
message for source code that uses this standard API.

Change-Id: Id8ed5e6697f9fbe0eecdaf8317a109e6ca45225d