Dispose Link instances when they lose all their connections.
When a Link instance loses all connections, we delete it from the links_
collection of the StoryImpl.

At teardown of StoryImpl we used to write link data of all links before
deleting the links. This would be subtle to maintain since now link
instances can be deleted at any time. However, we can drop the
last write to a link, because all values are always written
when they are set, so the last write is redundant. This makes
the tear down of links synchronous, and there
is no race between links being written to the ledger and links
being removed because they are disconnected.

Change-Id: I6d26ba5c1cc4337cd765debaeab2be42c3fcd5fd