Add Ready notifications to ModuleWatcher and StoryWatcher.
Change the Watcher interfaces of Module and Story to
a single OnStateChange() method that takes an enum indicating
the new state, rather than one method for every state.

Remove the OnData() method from the StoryWatcher interface.
The same notification can be obtained through the LinkWatcher
on the root link.

Collateral cleanup:

* Use the new Binding.NewBinding() method in a few places.

* Rename StoryState.NEW to StoryState.INITIAL lest it creates
  conflicts in the dart language binding. (Dart uses lowercase
  enum values, and new is a keyword.)

FW-71 #done

Change-Id: I2f8c2cfac27c1f3a16d089476cb3e78f6059e40a