[devmgr][fs] Prevent lookup from returning positive status
The vnode op "lookup" currently cannot return a positive return value
(indicating a remote handle) for either MinFS or MemFS, but the
surrounding code halfheartedly attempts to deal with this impossible
situation. This led to the Clang Static Analyzer finding a bug in the
following steps:

- vfs_rename is called
- vfs_walk(vndir, &oldparent, oldpath, &oldpath) returns a positive
  value, but leaves "oldparent" uninitialized
- vfs_walk(vndir, &newparent, newpath, &newpath) fails, and oldparent is
  released without being initialized.

Step 2 is only possible if vn->ops->lookup can return a postive value
(indicating a remote handle), but it cannot. MemFS's lookup calls
"dn_lookup", both of which are unaware of remote filesystems.

To prevent this bug from triggering in the future (and to satisfy the
static analyzer) this patch adds assertions that a non-positive return
value is propagating through these steps.

Change-Id: Ieaf8b3e5ea864204f728baa16b894b93431e3948