Fix persistence of root data in CreateStoryWithInfo().
If a story is created by CreateStoryWithInfo(), then the StoryController
is thrown away, then recreated, and the story then started, the root link
data passed in CreateStoryWithInfo() are lost because they are only written
to the Link (and thus persisted) when the story is started.

We change this to start the story runner and create the root link as soon
as we access it. This changes life cycle of the story runner inside the
story controller quite a bit, but was desirable anyway.

We also ensure that the root link gets written if the StoryController is
thrown away, that the StoryController does not get thrown away while
a request is ongoing, and that the story is not started when a stop
request is still in flight. We apply the usual pattern of a request
queue bottleneck to accomlish this.

Change-Id: Ib10d70ddf2972988944f00a4d858b516cebf5a47