[syscalls] support more syscall attributes
Before, the syscall parse engine pretended to understand syscall attributes
but in practice is better to not care and just pass-thru the attributes
as strings so the generators do what they please.

In other words the attributes go from an enum to a std::vector<string>

Along with this comes a relaxation of the syntax. Before you can only
have two attributes, now you can specify any number of them like this
    syscall <name> attr1 attr2 ... attrn (args) returns (arg)

This in turn allows us to fix a bug in the current syscalls.inc which
has the wrong attribute. Now we differenciate between "vdsocall" which
has no gccc __attribute__() associated and "leafc" which does.

Change-Id: I91e804015f3ff93c0b5abbebf352e9428a4fced5