[dash] Fix argument joining
strncat concatenates C strings. This means it searches for the end of
the destination, interpreted as a C string. Since we concatenating
many things into a stack buffer, into a non-terminated dmctl string,
this is wrong. Since we know where in the buffer we want to write,
just strncpy there.

Change-Id: Id50576c6347c974fbef7d34dac4c34cb2330873a