[make][warnings] remove -Wnull-dereference
This seems to be more trouble than it's worth, since it warns in all
cases where any pointer can be dereferenced that the compiler doesn't know
if it's null or not. Especially shows up around uses of smart pointers and
container classes, since the compiler can't guarantee positively that pointers
are null or not, even if they absolutely can't be.

Also tends to not show up on debug builds because of various DEBUG_ASSERTs
that serve as the null check to satisfy the compiler. Since it shows up on
DEBUG=0 builds it mostly just causes headaches for folks.

Change-Id: I558c9f9d10888f3bf69b5094d5288a15a7047ba8