[futex] Simplify handling of killing threads by using an interruptable wait
Set the thread's "interruptable" flag when waiting on a futex using
wait_queue_block().  This means we no longer need WakeKilledThread().

The code for using "interruptable" is similar to the code in
event_wait_timeout() in kernel/kernel/event.c.

After this change, a thread that's being killed (and that's waiting on
a futex) will wake and remove itself from the futex wait queue, rather
than WakeKilledThread() removing the thread from the futex wait queue
when waking the thread.

We can similarly remove FutexContext::WakeAll() (which was unnecessary
since the change that added WakeKilledThread()).

I think this change fixes a race condition: Before this change,
there's a race if FutexWait() happens concurrently with the thread
being killed.  It's possible for FutexWait() to block the thread on
the wait_queue after WakeKilledThread() or WakeAll() has completed.
If that happens, the wakeup will get lost, and the thread termination
won't complete.  After this change, the new check+wait in
BlockThread() should not have that problem.

Change-Id: Iff74d04b792628ca59f64c59b1968c08207272ae