[futex] Remove unnecessary UserThread::state() check
This check was present to handle the case where a thread/process is
killed while a FutexWait() syscall is in progress.  (However, I don't
think it worked fully because the thread/process could be killed after
the UserThread::state() check.)

The case of the thread being killed is now handled in
FutexNode::BlockThread().

Note that this access to UserThread::state_ was racy -- it was done
without holding state_lock_ so technically that could lead to
undefined behaviour.  We remove the state() method to discourage racy
uses in the future and because it's now unused.

Change-Id: I1295d87266cac40a795875b514b77b0351e5f69d