apps/netconnector: allow apps to register as responding services
We don't currently have any way for netconnector to launch a responding
service in a user context unless netconnector itself is running in that
context. Until we have a solution, we can address this issue by allowing
a running app to register with netconnector as a responding service. A
user can launch a responding service in a user context from the command
line. That service can register with netconnector, which runs in the
'boot' context. This CL provides support for that. It also factors the
service definitions into the core NetConnector functionality and a
'NetConnectorAdmin' administrative interface.

Change-Id: I6ef3781d4a46640068e0f06842a8602b5e84424d