[devmgr][fs] Remove V_FLAG_REMOTE
This flag is redundant; the state is expresses is encapsulated within
vn->remote. In fact, storing the singular state in multiple locations
can be the source of bugs: if "V_FLAG_REMOTE" is set, but "vn->remote"
is not (or vice versa), then the vnode would be in a corrupted hybrid
state.

Change-Id: I597ce32f95c5400fbdd8afa6e26a3c6efcf221b3