Fix ModuleControllerImpl::TearDown after https://fuchsia-review.googlesource.com/#/c/17674/.
StoryImpl::Stop() relies on the module being removed *before* it invokes
its TearDown() callback, because it continues only when all module
connections are disposed, but the change above changed that order,
because if the module controller deletes itself before invoking te
callbacks, callbacks from fidl methods on the module controller instance
are never transmitted.
The solution is to split removing the module controller from the story
and actually deleting it, so the two parts can happen in different order
from invoking teardown callbacks.

Change-Id: I85311c54bb0ebc303bb584a056579036ff33f0a5