[uapp][installer] Tweak how unmount_all() works
Add a pause after unmounting so devices become free.

Don't treat umount failures as terminal. If nothing is mounted
at a path ERR_BAD_STATE is returned. This error is not specific
enough to know what to do. Anyway, the proof is in the pudding,
can we open the device or not?

Change-Id: Ie542fd242704a6fcd930dd45d7cb3499bab1a0b7