[ddk] ddk/binding.h: Avoid designated initializers for C++ compatibility
This makes the MAGENTA_DRIVER_BEGIN/END macros work in C++ with GCC.

Though designated initializers aren't in standard C++, GCC and Clang
both support them in C++ as an extension.  However, GCC's support is
less complete than Clang's.  If we attempt to initialize .driver.name
with a designated initializer, GCC complains with:

  error: C99 designator ‘name’ outside aggregate initializer

(GCC doesn't like initializing a char array with a string literal
that's shorter than the array size when this is done via a designated
initializer.)

For GCC, it would work to omit the field names only inside .driver,
but it's more consistent and C++-standards-friendly to omit them for
all of the fields here.

Note that we can't do these particular initializations at run time
because this data structure can be read via a PT_NOTE segment (by
uapp/driverinfo/) without running any code from the DSO.

Change-Id: Icdabf88a796152c04add8e3a4f93f46ccb4b9d32