[unittest] Fix unittests for C++
Capturing references (auto&) for the LHS and RHS in the UT_CMP macro
can cause problems when the expression is a function which returns a
reference (like mxtl::min).  Capture values instead (which MAtches the
C macro expansion behavior)

Also, ASSERT_(NULL|NONNULL) were not returning from the test when the
failed like the should have.  Fix this.

Change-Id: I3facf6df82f7a42f93e297fb7e2249b9f126da77