[ddk] ddk/binding.h: Make use of attributes work in C++ with GCC
GCC has a quirk about how '__attribute__((visibility("default")))' is
interpreted for const variables in C++.  See the comment in the code
for more explanation.

This applies in GCC 6.2.0 at least.  Clang does not have this same
issue.

This quirk is probably related to the way const variables are treated
differently in C++ compared with C.  In C++, when a const variable is
defined (e.g. "const int foo = 1;"), the variable is not exported from
the compilation unit unless there's an "extern" declaration for it.

Change-Id: I315c637469e2feba0c12e53e2d39d15d8f4acce9