[minfs] Fix rename leak
Previously, on rename, the target vnode was overwritten in the
containing directory, but the inode itself never received a decremented
link_count. As a consequence, if "rename(foo, bar)" was called, 'bar'
would be made inaccessible, but it would not be deleted. This behavior
leaked both an inode (for bar) and data blocks (for any blocks held by
bar).

This patch fixes this issue and slightly refactors renaming.

Fixes: MG-449

Change-Id: I3a2509fc64b86bb6828582d266075f2b13f06ed2