[ddk] Update mac addr ioctl for using an array
The _OUT wrappers assume the output size is the size of the output
type. For mac addresses, we really want an array of such things, so
we require the caller to pass in the size of the array by using _VAROUT.
Implementations of this ioctl will check the length to ensure that the
result fits.

Change-Id: If276dfcab9ec864e8762c36c83eb59b0f1ae6234