Merge story_runner application into user_runner.
Run story runner in the user runner process, it doesn't need to
be its own process. This removes some internal fidl interfaces.

We still preserve the separation of a StoryImpl and a StoryControllerImpl
class, even though they have closely related life cycles. They will be
merged or factored separately in a followup.

One subtlety is that we now have asynchronous operations that are not
fidl method invocations (we had some in StoryProviderImpl before). The
problem there is that if a callback from such an operation deletes
its target this has much more complicated implications than in a fidl
call, where both callbacks and anything that would delete a service
instance are always invoked through a runloop and thus it's impossible
to delete a service instance from inside the callback of its methods.

Several approaches to deal with this problem were tried during the evolution
of this change. The change now works, and I think we have found something
close to optimal. For the record, a few notes about the problem and its
solutions:

* The problem is that when a callback from one method wants to delete
  the target of the method (for something like a Stop() method after
  which its target is no longer needed) or a related instance that
  is not owned by the target but was provided to it, then after the
  callback returns, as well as all other functions up the call stack,
  |this| might be deleted for the remainder of the function.

* One approach is to ensure that every function that invokes callbacks
  is aware of the possibility and does not acces this after invoking
  callbacks. This is complicated because it doesn't happen automatically,
  thus needs to be built explicitly. Also, it's not necessary in every
  instance, but this distinction may be made a few levels down the call
  (up the call-back) stack, so effectively it must be applied everywhere
  even where superficially it's not necessary. (In the change, the
  problem arises from the result callback of a Sync() call that is
  invoked for Story teardown.)

* The other approach is to push callbacks onto a run loop, and thus avoid
  to expose the remainder of callback invoking functions to potentally
  deleted this. This always happens automatically for result callbacks of FIDL
  methods, and thus is needed to do explicitly only for result callbacks of
  non-FIDL methods. That's easier and more local to decide is necessary
  than the previous remedy. This is the solution we used, see StoryImpl::Stop().

* A third approach would be to give such objects ownership to themselves and
  let these objects delete themselves once they know they are done, so that
  their clients are not tempted to clean them up in a result callback. This
  is what FIDL service implementations do on disconnect.

Collateral fix: Link data that has changed in the ledger was not
previously applied to the story, because no watcher was registered.
It now is.

Collateral fix: If a module is unlinked, the stop callback is still
invoked through the run loop, albeit without a timeout. This is more
uniform.

Change-Id: I08a15302adec8038a7530d46a248d5228d75fd33