Introduce the ByteBuffer abstraction
Added the "common" directory for common packet manipulation and testing
helpers. This CL introduces the ByteBuffer interface and two
specializations thereof where one is stack allocated and the other is
based on malloc.

The general idea is here is that transport and packet layers don't need
to know how a packet is allocated and a buffer interface should abstract
away its allocation, move, and copy semantics. In the future we can add
various specializations of this, e.g. using a malloc or VMO backed slab
allocator.

This CL also adds a top-level BUILD.gn defining the high-level build
groups.

Change-Id: I93b01625e3bdaac0a9754c2cabc86e29dee2c660