[aslr-analysis] Bring up to speed with system changes
Now this makes use of jobs and handles a change in the behavior of
program loading around relative paths.  Previously using launchpad
to launch "aslr-analsysis" would work, but this apparently changed.
Additionally, this fixes a use-after-free issue with a handle,
caused by a change in launchpad behavior that landed while the
original patch was out for review.

Change-Id: I2fb616aafa7f313f8336e1ebeef242840b31dbfe