[musl] Add clang thread capability("mutex") annotation to libc mtx_t type
This adds the annotation __attribute__((capability("mutex"))) to the C11
standard library type mtx_t so code using clang's -Wthread-safety
annotations can annotate its manipluation of mtx_t types directly. This
is analogous to the annotations for libc++'s <mutex> types:

https://fuchsia.googlesource.com/third_party/libcxx/+/master/include/__mutex_base

and should be no-ops for code that is not using clang or not using
thread-safety annotations. Even for code that is using clang and
thread-safety annotations, this annotation by itself should not cause
issues. Having the annotation on mtx_t allows adding other annotations
to other code that wants to use mtx_t but does not impact code that
uses mtx_t without annotations. libc++'s annotations are guarded by a
_LIBCPP_HAS_THREAD_SAFETY_ANNOTATIONS feature guard. The annotations in
this patch aren't guarded, but we could introduce a guard if decided.

The main risk of incompatibility is that we write code that expects
mtx_t to be annotated and then use that code with a libc lacking the
annotations. This code would likely fail to compile with -Wthread-safety
enabled unless the annotations referencing mtx_t were guarded.

Without this annotation, mxtl's AutoLock implementation in the
 #defined(_KERNEL) case fails to compile.

Change-Id: I16d5819818f512889a46f88876778763d96cc15d