[minfs][fs] Fix MinFS Block bug from C++ conversion
During the C --> C++ conversion, somehow, "kMinfsBlockSize" was used
where "kMinfsBlockBits" was intended. This resulted in unexpected
results for any bitmap access (inode or blockmap) greater than "8192".
Since most of our filesystems tests are small, this was not caught
during testing -- however, it did cause the maxfile test to fail, but it
was not seen since "test failures" just resulted in a different print
message (PASSED vs FAILED) with no summary at the end.

This patch does the following:
 - Fixes the MinFS bug
 - Strengthens the Maxfile test to find the bug earlier
 - Updates the filesystem test suite to "exit on error" faster, so these
   errors can be caught more quickly. I can update this to something
   more fancy later, but I'd rather keep the test suite "stupid and
   obvious" until it's actually converted to a utest.

Change-Id: I1be4823f892da7e961282446dbbc99cc5818fd15