[sysgen] Clean up generated function attributes
The important distinction of cases like num_cpus is 'const',
not 'leaf'.  So call the attribute in .sysgen files 'const'.

The 'leaf' attribute's semantics actually apply to all vDSO
entry points.  So generate that attribute on every function
declaration.

Also, use __*__ names for non-keyword attributes so the C header
is not sensitive to random user use of identifiers as macros.

Change-Id: I1fd3f4b597e8314cc63e402142273f0968d68ce5