[utest][debugger][exception] Use a volatile atomic for the watchdog flag
A flag set in another thread should be tested using proper atomic
access.  Adding volatile ensures that the compiler doesn't decide
it knows what can and can't touch the flag variable and thus that
it can hoist the test outside the loop that checks periodically
if another thread has set the flag.

Change-Id: I33b66c6f802026e56a7f2b48ff24baf1a65413f6