[bootloader] Support the headless mode
When there's no display connected, there is no framebuffer. Magenta
seems to be handling that case, but our bootloader doesn't. This
change adds a couple of extra sanity checks to make sure we don't
die because of null dereference in that case, which seems to be
enough to get the bootloader working in the headless mode. However,
I haven't through the entire codebase and checked every codepath,
so there might still be cases when that might happen.

Change-Id: Id9679a5d10346d335486bc9dbac985bd2ff53da7