[ulib][musl][ldso] Fix x86_64 stack alignment in early entry code
The assembly code for the dynamic linker's startup was failing to
align the stack pointer correctly before calling into C code.  Clang
generates code using SSE instructions that are specified to fault if
the stack is not properly aligned.  But QEMU does not enforce this
alignment requirement, so the bug only bites on real hardware.

Change-Id: I619039c7e0dd5cf2e1bc341eed3f8842f1bffb05