[vmar] Get rid of MX_VM_FLAG_ALLOC_BASE and friends
This workaround is no longer needed due to Roland's work
with integrating VMARs into the loader.

Note that there is an erroneous comment in the deleted code saying
to remove it once the DDK is migrated to the new interface.  That
comment was incorrect.  The only dependency was the loader.

Change-Id: I6eb2a79747b4c9c7e32a860bf89de67a0992c5d2