[docs] Slightly clarify what annotations should be on the declaration
The wording for which annotations should be on a function's declaration
vs its definition was slightly confusing. This tries to clarify that
annotations describing part of a function's interface should be on the
declaration. The text in the Limitations section below states that
TA_NO_THREAD_ANALYSIS should be on definitions, but this wasn't clear.

Change-Id: Idc2790d7a750c702cc6b8964141b75ce3d59e0a4