[syscalls][channel] fix race condition in cancelation
If the Call is canceled before it waits, it will always receive
NO_ERROR as the wait result, and be unable to differentiate between
a success and a cancelation.

We upgrade the bool in MessageWaiter to a mx_status_t so we can
know exactly why we were notified and handle it accordingly.

MG-473 is tracking the limitation of the event_signal_etc() API
which this change is working around.

Change-Id: I146e69c69cf46f338ecc60e6fb9d6658e002bf29