[devhost] return correct error when iostate is stale
"stale" iostates happen when a device is removed while an RPC is
in flight.  We were correctly refusing to pass the RPC on to
device ops in this case, but incorrectly returning NO_ERROR.

This then sometimes caused attempts to OPEN devices that were
in the process of being closed to appear to return success but
not return any handles (a strange and invalid response).

We now return ERR_REMOTE_CLOSED, to better represent the
"you can't talk to this device any longer" state.

Change-Id: I9ad02f88023ba5abc090ae6733c352de207547fa