[fs] Fix Unmounting via channel_call
As mentioned in the code, we don't care about the reponse from the
filesystem server (at the moment) -- we merely care that the filesystem
received the 'unmount' command, and had the opportunity to shut itself
down.

This change means that if the filesystem shuts itself down before
responding to 'mx_channel_call', we consider that a 'success' case,
rather than a failure case.

Change-Id: Ia185e96f8f5fe1600f796ab109dfd52dc1f89824