Remove the modular resolver.
Resolver will be provided by TQi directly to modules. It might be
that we do something alternative to resolve modules in the story
runner too that requires other input from the caller, for example
instantiating modules from components rather than applications,
with more variability as to cardinality and isolation, but this
is likely to look different from thej dummy resolver we had until
now.

Change-Id: Id1bb8e874e4f64173e86dae2adf4775eac6e420e