apps/media: fix issue with problem reporting
The Problem struct has an optional 'details' field, which should be null
if no details are provided. The C++ code that generates Problem structs
uses empty strings instead, producing details fields with empty strings.
This CL addresses that in the demux component. It also removes an unused
problem reporting method in the source component.

Change-Id: Id91739ad40233b71c773e4f48b8f4c51043f0616