[mxtl] Put memory allocation of Slabs behind an abstraction layer.
The way in which SlabAllocator<>s obtain memory for their internal
slabs depends on whether or not the allocator is running in the
kernel, or running in user-mode.  In user mode, the posix
aligned_alloc is used.  Currently the kernel does not implement this,
so malloc along with an alignement assert is used instead.  Moving
forward, even malloc is going to disappear from the kernel, and this
implementation will be replace with something more sophisticated and
appropriate for the kernel's needs.

Change-Id: I3d9d4b1ecf7eeb33f62f35af5724141b149447ef