[ulib][musl] Avoid ARM64 autovectorization in memset as a workaround
See MG-480 and MG-481 for the real bugs.  This works around those bugs
by preventing the LLVM code generator from producing the (valid, and
well-optimized) code that turns those latent bugs into active crashers
on ARM64 hardware.  With this workaround, the system boots and passes
tests on Raspberry Pi 3 just as well when built with Clang as with GCC.

MG-480 # workaround
MG-481 # workaround

Change-Id: I7d150a610d4b80a31164f4a0b75240985ac2e366