Add push-package.py
Added a simple Python script that looks at a given package file from
packages/gn and copies the enumerated contents from the build output
directory to the appropriate destination paths under /tmp using netcp.

This script is dumb in that it does not figure out dependencies from
other packages AND it copies everything every time it gets run, i.e.
it doesn't have the smarts to incrementally push the subset of files
that changed.

Change-Id: Ie0627a5db689380e5ba9dcfff4de59fd77d0a633