[kernel][msm-trapper][clang] fix inline asm for gic accessors on clang
Apparently the MSR/MRS instructions need to use the full 64bit registers,
which gcc happily does (even if you feed it 32bit args), but clang complains
that the sizes dont match up, which seems fairly correct.

Cast things around to make it line up with the registers size that it has
to use.

Change-Id: I5851f7fee3e760c47310d050db61d2c086d940d2