[magenta] Remove ProcessDispatcher::thread_list_lock_, always use state_lock_
The set of threads in a ProcessDispatcher can only be mutated while
holding state_lock_ as changing the number of threads in a process may
change its state (i.e. exiting the last thread). This removes the
thread_list_lock_ which was acquires with the state_lock_ held in all
cases except for lookups, which are expected to be less common.

Change-Id: Ic4a351ea3e044e597718314b83691b8f905fd161