apps/media: add MediaPlayer widget to lib/flutter
This CL adds a MediaPlayer widget that can be used to play audio-only or
A/V content. The widget is the 'view' for MediaPlayerController, submitted
previously. The Widget integrates a play/pause button and a touchable
progress bar with duration and progress text. Other visuals are left to
the hosting UI, which can access the controller to control and interrogate
the underlying player.

Change-Id: I4795b94a29f9c044d6b05ce8e79ad98b70946f67