[kernel][magenta] Fix Kill for threads that were never Start-ed.
If a thread is created but never started (e.g.
test_thread_start_on_initial_thread), then the reference count for the
thread is never correctly reset. The AddRef call in Initialize is
never balanced by a Release call. This causes the UserThread to leak,
and to hold a reference to the related process, subsequently leaking
the ProcessDispatcher as well.

MG-446 #done

Change-Id: Id622455211684ef56deb9628d3a7500487f93d10