[mxtl] Add MANUAL_DELETE behavior to SlabAllocator
Refactor slab allocator flavors so they are based on an enum instead
of just a bool used to determine static vs. instanced.

Introduce a new flavor of SlabAllocator<> called a MANUAL_DELETE
SlabAllocator.  MANUAL_DELETE SlabAllocators are instanced, but
objects which are SlabAllocated<> from a MANUAL_DELETE slab allocator
do not carry the extra overhead of a pointer used to return home
during execution of the SlabAllocated<> delete operator.  Instead, the
delete operator of the object is hidden (so users cannot accidentally
call it), and users must manually return their objects to the proper
allocator via an explicit call to the allocator's Delete method which
will run the object's destructor(s) and then return the memory to the
allocator's free list.

Change-Id: If8908d6b7c04d5663c8f6cbc031d413c8960bd99