[arm64] Process pending thread signals on sync exceptions
This brings the arm64 code in line with the x86 code, and corrects an
issue in which killing a thread that is suspended due to pending
exception processing (for e.g. fatal page faults) would never succeed
due to the thread getting stuck in a page fault loop and never
processing the kill signal.

MG-483 #done
MG-486 #done

Change-Id: Ice22d55f071c85f6698cc528318f0e69dae1fb1e