[uapp][installer] Allocate data partition space
Add a data partition if space is available to do so.

This also change the function that looks for available to space
in the case that a large enough unallocated section is not
found. In that case the function returns the largest available
section, in the event there is any free space. This allows the
caller to effectively request a size, but then settle for less.

Change-Id: I440ef7c5d1cd3a6a67ce687894a148055c145316