[devmgr][mxio] Add a system-wide loader service
Add a dmctl ioctl to get a handle to a shared system loader service,
which translates paths like "libhid.so" into VMOs containing the named
object's data.

Launchpad and other users of mxio_loader_service() will automatically
prefer the system loader service if available.

Before this, loading was done by in-process threads, which meant that
those processes needed filesystem access. This is a step towards
supporting dynamic binaries with no filesystem rights.

For now, the system loader service uses the same code as the in-process
loader service, but this may change. It also needs to be smarter to
avoid starvation from bad processes (MG-491).

Moving the loader-related prototypes to a new header caused some ripple
effects for code that depended on transitive includes from mxio/util.h.

For MG-451: "loader service"

Change-Id: I8adb534b55f96334017e979da6af263e19095d84