[driver][virtio][arm64] fail without an assert on arm64 on block devices
At the moment the pci bus will not properly map in PIO mmio area for arm64
machines, which causes pci->map_mmio to return a negative handle, which libmx
currently interprets as a 'good' handle. Change the test to test for negative
until system handles are made unsigned, and pci->map_mmio api is changed to match.

Change-Id: I3aad0e149638999184a187c9c957851710f24ccf