Move state from global variables to local variables and properties.
With this, most state is only acesses from code that is there only
when the state exists, so fewer checks are necessary. Also life cycle
is more explicitly aligned.

The state of the application is kept in the State<> of a stateful
widget at the top of the view.

However, the state object itself is flutter independent and holds
the app context and the module, as well as provides a way to invoke
notification callbacks on model changes. The former gets instantiated
in main and passed into the top level widget of the flutter application.
The actual flutter State object gets then created by createState() and
wraps the pure state object.

This keeps state local but generalizes better to multiple flutter state
instances in the application: If the application would have multiple
stateful top level widgets (under different routes, for example), each
would create its own flutter state instance, but the pure state instance
would be shared between the flutter state instances.

This also drops use of global keys, because all state is now accessible
locally.

Change-Id: I04cd60ed37c1b6be06e2e1c8202e702bdf4dbed7