[dash] Add a shell controller protocol.
This patch adds a simple controller protocol, allowing to exchange
control messages with dash by passing a channel handle as
MX_HND_TYPE_USER1.

The protocol is described in `controller.h`, and currently only includes
commands sent from the shell to the controller:

 - "get_history" retrieves the shell history.
   response payload: empty
   response handles: a vmo where shell history is stored as '\n'
     separated entries, including a trailing '\n' after the last entry.
 - "add_to_history:<entry>" adds the given <entry> to the shell
   history.

Together, these allow the controller to manage the shell history (for
example, for persistance and synchronization).

Change-Id: I7c1d603d69e642a2140609c5718008d410c66d52