[linenoise] Fix crash that can occur if another process writes to the terminal
Before this change, if the prompt was wider that the terminal,
refreshSingleLine()'s logic could pass a negative size to abAppend(),
causing a crash in memcpy().

I found that could get triggered if I added some noisy logging to the
kernel.  That would race with linenoise's getColumns(), which would
wrongly think that the terminal width is 1.  (This was when using the
serial console under QEMU.)

Fix refreshSingleLine() to handle this case, and add
assertions/comments about the other parts.

Change-Id: I166567c2c401b6559980dde4cc9f168bea2e4d68