[channels] fix message queue teardown
A message could arrive from the other endpoint after the message
queue was discarded in on_zero_handles() and before the assert
in the destructor.

Move message queue teardown to destructor, at which point there
are no other references to this endpoint remaining and we can
be sure we're discarding all undelivered messages.

Previously we avoided message queue teardown in the destructor
to reduce impact of recursive destruction of channels with messages
with channels with messages... but recent changes avoid that
recursion, handing handles within messages to a dpc for teardown.

Resolves MG-493

Change-Id: I4b60370fe4f6ca94a03a4ec987be6d8653834432