[kernel][timer] refactor and add a spinloop signal to a trylock in some timers
Refactored some code that was used in a few places to pull into the
timer sources. Add a spinloop pause as it goes around the loop waiting
for the spinlock or the timer to be canceled. Make sure the cancel logic
properly signals to wake up a loop.

This fixes a livelock in arm64 on qemu with the not-exactly-smp smp
emulation that causes a trylock spinner to never release the cpu.

Change-Id: I984e74febf44aebe035359b2d6b6124ad02d409f