[build][kernel] remove libgcc/compiler-rt from the link line for the kernel
Since the switch to hard 64bit we are using nothing out of these libraries,
and this avoids having to build -mno-red-zone versions of these libraries.

We may need to revisit this if we port to another architecture that needs it
but for now this should keep us honest about using slow math in the kernel.

Change-Id: I5c221d5e240e9d8e0106617564807255df641294