[dispatchers] Don't overload the GetDispatcher* variants
Having them be true overloads was a bit confusing, since they were
really doing different things: getting a dispatcher from a handle;
getting a dispatcher and rights from a handle; getting a dispatcher
and checking that it has certain rights.

This change also puts out paramaters last, to be consistent.

Change-Id: I6fd92647dd28d1347b85d1a862aa6980370236da