apps/media: support remote control in flutter example app
This CL adds the ability to remotely control a running media player using
the flutter media player. A new asset type ('remote') is supported, allowing
the user to connect to a remote media player and control it. We don't
currently have a way to invoke a media player remotely, so this only works
for controlling media players that have been started already.

Change-Id: Ida9ae58b970e17c8ceb0049ee6b22d496e413085