[musl] Add explanatory comments to pthread_cond_timedwait() and cnd_timedwait()
The algorithm used by Musl's condvar implementation is quite complex
and subtle, and the implementation is cryptic with little explanation.

This change adds some explanation of how the algorithm works.

cnd_timedwait() has a duplicate implementation, so I've copied the
comments there too, changing the pthread_*() references to refer to
mxr_mutex_*() and cnd_*() instead.

MG-361

Change-Id: I9023435f6a629a398ccfcf3729b7a9d865ed0a53