[common] Introduce MutableBufferView
This CL splits the BufferView class into const and non-const
classes. BufferView now provides a purely immutable view over a valid
piece of memory while enforcing const-correctness. MutableBufferView
is an alternative in which the originator intends for the contents of
the buffer to be modifiable.

The ByteBuffer interface has also been split into the ByteBuffer and
MutableByteBuffer classes to make the inheritance chain make more sense.

The API differences are:
   - The respective constructors provide a const-correct interface for
     initialization.

   - All non-const methods of BufferView fail fatally.

Change-Id: I84a3a09f88a9dfe8b75280d1038176e81864416e