[devmgr] Spawn thread to shutdown devmgr.
In the process of shutting down gracefully, devmgr must
 1) Send a signal to filesystems, telling them to shut down,
 2) Wait for a response or timeout,
 3) Actually turn off power via 'devmgr_poweroff' or 'devmgr_reboot'

However, there is a bit of an issue with steps (1) and (2): Some
processes (filesystems) have a dependency on the devmgr when shutting
down, and cannot complete process termination without interaction with
the devmgr. As a consequence, waiting in (2) prevents (1) from
completing.

To break this dependency, allow the "shutdown" commands to run
asynchronously in a background thread.

Change-Id: I865d2018c32198f74dc5ebc3f9e3429cca165734