Suggest shell example story.
A set of modules that allows to make
suggestion proposals from user input.

Introduces the concept of a ViewHost,
which is a module that only displays
UI from its child modules, without having
any UI on its own, like a View Controller.

Used here to split the story into a View
module that displays UI, and a View Controller
module that orchestrates the View and calls
services.

Change-Id: I110d4b6a1c03906c20292a415cffaf0badd9d37d