apps/media: add frame layout info to media type
Video frame layout information, specifically line stride and plane
offset, have heretofore been inferred from the pixel format and
'coded size' of the video. This was done with some common code that
embodied a bunch of ffmpeg-specific peculiarities. Now that we're
going to use libvpx for decoding VP8 and VP9, it should be up to
the decoder to determine frame layout, and it should be communicated
explicitly in the video media type. This moves the ffmpeg-specific
layout code to src/ffmpeg where it belongs and will allow libvpx to
use distinct layout logic. In the future, we may need to add the
option of specifying frame layout on a per-packet basis, but we'll
cross that bridge when we come to it.

Change-Id: I47650b5bd94f89626378010ea4da9a6073930595