[kernel][magenta] Ports V2 part 3
Given changes in state observer this CL features a great
simplification of the pointer and ownership. See
ascii art in the dispatcher header for more details.

The core of repeating signals is there but there are
a couple more details pending and the count is not
fully backed.

Along the way we now handle on_zero_handles() callback.

Change-Id: I5ac074a08980e87935097b724dc4926bcf7109da