[kernel][sched] first stab at dynamic priorities
Based on blocking, unblocking, or using up a thread's time slice,
dynamically adjust the priority of the thread by +/- a few levels.

In the initial implementation, allow the effective priority of a thread
vary by +/-4 priorities, with the default user priority set to 16.

If a thread runs to to time slice expiration (currently 50ms) the thread
is deboosted by 1 all the way to -4. If it simply reschedules it will
deboost by 1.

When a thread is woken up it is boosted by 1, to 4.

This seems to improve system responsiveness, especially with background
threads running continually.

Change-Id: I47185ac2b42489421a56419c4b22c79e8da541cd