Use app namespace instead of modular for public interfaces
This moves the C++ and FIDL interfaces for the application model from
the 'modular' namespace to the 'app' namespace to more clearly separate
them from other parts of what's currenty called modular. To ease the
transition, I've left C++ aliases for the library types in the modular
namespace that it's possible to update C++ only callers in smaller
portions instead of doing them all at once. There's no equivalent
aliasing mechanism for FIDL, so all the callers have to be updated
simultaneously with this patch (there's not that much).

Change-Id: Id47900b57722acb76962efa9449587f87a27d81e