MessageQueues: Initial implementation of ComponentContext.{*MessageQueue}() and a MessageSender.
The message data are not persistent.

In this CL:
- MessageQueueManager: stores mappings of [component id -> queues], and
[tokens -> (component id, queue)]. Actual message queue implementation
is in MessageQueueStorage.
- MessageQueueStorage: stores messages, and has implementation for
MessageQueue.Receive() and MessageSender.Send(). Implements the
MessageSender interface, and manages a MessageQueueConnection.
- MessageQueueConnection implements MessageQueue and forwards calls
right back into MessageQueueStorage. It exists to mostly signal when a
|MessageQueue| interface is closed on the other side.

Some fixing to do:
- Clean up the plumbing for making ComponentContextImpl (maybe a
ComponentContextFactory?) -- right now, AgentRunner and
MessageQueueManager are plumbed out in two places solely to initialize
ComponentContextImpl.
- Rationalize token strings.
- Possibly remove MessageQueueConnection and instead extend BindingSet
to signal connection closures (not sure if worth it though).

Change-Id: I4e474c963e5e12c4953f927c269424595eeaf38d