[devmgr][cleanup] call this agent "dev coordinator"
It lives in devmgr (but may move some day), and handles requests
from devhosts (but is not a devhost).  It coordinates the top level
tree of devices and arranges to publish them in devfs, so let's call
it coordinator to keep it distinct from other services and components
in devmgr and devhosts.

Change-Id: I48312fd74f11fe07e43b3983db10eb03800fdff3