[magenta][hypervisor] Add beginnings of a hypervisor.
This does the very basic task of enabling VMX on each CPU. Next step is to
setup MSRs and to launch into a VM.

I'd like some early feedback before I dive too deep. A lot of this is subject
to change as we learn more, and iterate on the implementation.

Change-Id: Idce0d36d73e9a3be8e607f6c9b1a29597cfc4f51
NOTE: Only Intel CPUs are currently supported.